Add explicit short/int casts to (Big|Little)(Short|Long)
Make sure the signedness stays the same after conversion

+++ Quake/common.h
@@ -177,10 +177,10 @@ void Vec_Free (void **pvec);
 #define host_bigendian 0
 #endif
 
-#define BigShort(s)    SDL_SwapBE16((s))
-#define LittleShort(s) SDL_SwapLE16((s))
-#define BigLong(l)     SDL_SwapBE32((l))
-#define LittleLong(l)  SDL_SwapLE32((l))
+#define BigShort(s)    ((short)SDL_SwapBE16((s)))
+#define LittleShort(s) ((short)SDL_SwapLE16((s)))
+#define BigLong(l)     ((int)SDL_SwapBE32((l)))
+#define LittleLong(l)  ((int)SDL_SwapLE32((l)))
 #define BigFloat(f)    SDL_SwapFloatBE((f))
 #define LittleFloat(f) SDL_SwapFloatLE((f))
